backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.12.4
pydantic-settings==2.1.0
email-validator>=2.1.1
psycopg2-binary==2.9.11
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx<0.28

